+++ django_mapengine/views.py
@@ -57,6 +57,7 @@ class MapEngineMixin(ContextMixin):
             "regions": settings.MAP_ENGINE_REGIONS,
             "result_views": {},  # Placeholder for already downloaded results (used in results.js)
             "zoom_levels": settings.MAP_ENGINE_ZOOM_LEVELS,
+            "cluster_layers": [cluster.layer_id for cluster in settings.MAP_ENGINE_API_CLUSTERS],
             "choropleths": {choropleth.name: choropleth.as_dict() for choropleth in settings.MAP_ENGINE_CHOROPLETHS},
             "layer_switch_class": settings.MAP_ENGINE_LAYER_SWITCH_CLASS,
         }