+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 
 Here is a template for new release sections
 
-## Unreleased
+## [Unreleased]
 ### Added
 - error message if region zoom level does not work with distill zoom level